[docs] : 컴포넌트 구성 수정
