--- src/components/mypage/SubCategoryList.jsx
@@ -1,4 +1,0 @@
-const SubCateGory = () =>{
-
-    
-}
